v3.0.0 updates
