+++ README.md
@@ -2,9 +2,9 @@
 
 > For educational, authorized and/or research purposes only.
 
-o365spray a username enumeration and password spraying tool aimed at Microsoft Office 365 (O365). This tool reimplements a collection of enumeration and spray techniques researched and identified by those mentioned in [Acknowledgments](#Acknowledgments).
+o365spray ia a username enumeration and password spraying tool aimed at Microsoft Office 365 (O365). This tool reimplements a collection of enumeration and spray techniques researched and identified by those mentioned in [Acknowledgments](#Acknowledgments).
 
-> WARNING: The oAuth2 module for user enumeration is performed by submitting a single authentication attempt per user. If the module is run in conjunction with password spraying in a single execution, o365spray will automatically reset the lockout timer prior to performing the password spray -- if enumeration is run alone, the user should be aware of how many and when each authentication attempt was made and manually reset the lockout timer before performing any password spraying.
+> WARNING: The Autologon, oAuth2, and RST user enumeration modules work by submitting a single authentication attempt per user. If the modules are run in conjunction with password spraying in a single execution, o365spray will automatically reset the lockout timer prior to performing the password spray -- if enumeration is run alone, the user should be aware of how many and when each authentication attempt was made and manually reset the lockout timer before performing any password spraying.
 
 > If any bugs/errors are encountered, please open an Issue with the details (or a Pull Request with the proposed fix). See the [section below](#using-previous-versions) for more information about using previous versions.
 
@@ -29,7 +29,7 @@ usage: o365spray [-h] [-d DOMAIN] [--validate] [--enum] [--spray]
                  [--timeout TIMEOUT] [--proxy PROXY] [--output OUTPUT]
                  [-v] [--debug]
 
-o365spray | Microsoft O365 User Enumerator and Password Sprayer -- v2.0.4
+o365spray | Microsoft O365 User Enumerator and Password Sprayer -- v3.0.0
 
 optional arguments:
 
@@ -67,12 +67,13 @@ optional arguments:
   -l LOCKOUT, --lockout LOCKOUT
                         Lockout policy's reset time (in minutes). Default: 15 minutes
 
-  --enum-module {office,onedrive,oauth2}
-                        Specify which enumeration module to run.
-                        Default: office
+  --validate-module     Specify which valiadtion module to run.
+                        Default: getuserrealm
 
-  --spray-module {oauth2,activesync,autodiscover,reporting,adfs}
-                        Specify which password spraying module to run.
+  --enum-module         Specify which enumeration module to run.
+                        Default: oauth2
+
+  --spray-module        Specify which password spraying module to run.
                         Default: oauth2
 
   --adfs-url ADFS_URL   AuthURL of the target domain's ADFS login page for password
@@ -106,72 +107,28 @@ optional arguments:
 
 ## Modules
 
-o365spray has been packaged to allow for use within automation scenarios. If domain validation, user enumeration, or password spraying is a part of your proposed attack/recon automation, see the below modules and import usage exmaples.
-
 ### Validation
-* getuserrealm
-* openid-config -- *Currently Disabled*
-
-The validator can be imported and used via:
-```python
-from o365spray.core import Validator
-v = Validator()
-valid, adfs_url = v.validate('domain.com')
-```
+* getuserrealm (default)
 
 ### Enumeration
+* autologon
+* oauth2 (default)
 * office
-* oauth2
 * onedrive
-* activesync -- *Currently Disabled*
-* autodiscover -- *Currently Disabled*
+* rst
 
 > The onedrive module relies on the target user(s) having previously logged into OneDrive. If a valid user has not yet used OneDrive, their account will show as 'invalid'.
 
-The enumerator can be imported and used via:
-```python
-from o365spray.core import Enumerator
-loop = asyncio.get_event_loop()
-e = Enumerator(loop, writer=False)
-loop.run_until_complete(
-    e.run(
-        userlist,
-        password,
-        domain,
-        module,
-    )
-)
-loop.run_until_complete()
-loop.close()
-list_of_valid_users = e.VALID_ACCOUNTS
-```
-
 ### Spraying
-* oauth2
 * activesync
-* autodiscover
-* reporting
 * adfs
+* autodiscover
+* autologon
+* oauth2 (default)
+* reporting
+* rst
 
-> The oAuth2 module can be used for federated spraying, but it should be noted that this will work ONLY when the target has enabled password synchronization - otherwise authentication will always fail. The default mechanic is to default to the 'adfs' module when federation is identified.
-
-The sprayer can be imported and used via:
-```python
-from o365spray.core import Sprayer
-loop = asyncio.get_event_loop()
-s = Sprayer(loop, writer=False)
-loop.run_until_complete(
-    s.run(
-        password,
-        domain,
-        module,
-        userlist,
-    )
-)
-loop.run_until_complete()
-loop.close()
-list_of_valid_creds = s.VALID_CREDENTIALS
-```
+> The oAuth2 module can be used for federated spraying, but it should be noted that this will ONLY work when the target tenant has enabled password synchronization - otherwise authentication will always fail. The default mechanic is to default to the 'adfs' module when federation is identified.
 
 ## Omnispray
 
@@ -188,15 +145,20 @@ The o365spray framework has been ported to a new tool: [Omnispray](https://githu
 | [byt3bl33d3r](https://github.com/byt3bl33d3r) | MSOLSpray: Python reimplementation | [Gist](https://gist.github.com/byt3bl33d3r/19a48fff8fdc34cc1dd1f1d2807e1b7f) |
 | [nyxgeek](https://github.com/nyxgeek) | onedrive_user_enum: OneDrive user enumeration | [onedrive_user_enum](https://github.com/nyxgeek/onedrive_user_enum) / [blog post](https://www.trustedsec.com/blog/achieving-passive-user-enumeration-with-onedrive/) |
 | [Mr-Un1k0d3r](https://github.com/Mr-Un1k0d3r) | adfs-spray: ADFS password spraying | [adfs-spray](https://github.com/Mr-Un1k0d3r/RedTeamScripts/blob/master/adfs-spray.py) |
-| [Nestori Syynimaa](https://github.com/NestoriSyynimaa) | AADInternals: oAuth2 user enumeration | [AADInternals](https://github.com/Gerenios/AADInternals) |
+| [Nestori Syynimaa](https://github.com/NestoriSyynimaa) | AADInternals: oAuth2 and autologon modules | [AADInternals](https://github.com/Gerenios/AADInternals) |
 | [Daniel Chronlund](https://danielchronlund.com/) / [xFreed0m](https://github.com/xFreed0m) | Invoke-AzureAdPasswordSprayAttack / ADFSpray: Office 365 reporting API password spraying | [Invoke-AzureAdPasswordSprayAttack](https://danielchronlund.com/2020/03/17/azure-ad-password-spray-attacks-with-powershell-and-how-to-defend-your-tenant/) / [ADFSpray](https://github.com/xFreed0m/ADFSpray) |
+| [Optiv](https://github.com/optiv) (Several Authors) | Go365: RST user enumeration and password spraying module | [Go365](https://github.com/optiv/Go365) |
 | [byt3bl33d3r](https://github.com/byt3bl33d3r) | SprayingToolkit: Code references | [SprayingToolkit](https://github.com/byt3bl33d3r/SprayingToolkit/) |
 | [sensepost](https://github.com/sensepost) | ruler: Code references | [Ruler](https://github.com/sensepost/ruler/) |
 
 ## Using Previous Versions
 
-o365spray was recently rewritten (v2) and could have some hidden bugs remaining. If issues are encountered, try checking out the commit prior to the code rewrite (`v1.3.7`):
+o365spray has recently been rewritten and could have some hidden/unknown bugs. If issues are encountered, try checking out the commit prior to the code rewrites:
 
-```
+```bash
+# v1.3.7
 git checkout e235abdcebad61dbd2cde80974aca21ddb188704
+
+# v2.0.4
+git checkout a585432f269a8f527d61f064822bb08880c887ef
 ```
