+++ o365spray/__init__.py
@@ -1,4 +1,4 @@
-_V_MAJ = 2
+_V_MAJ = 3
 _V_MIN = 0
-_V_MNT = 4
+_V_MNT = 0
 __version__ = f"{_V_MAJ}.{_V_MIN}.{_V_MNT}"
