--- o365spray/core/__init__.py
@@ -1,10 +1,0 @@
-# Utils
-from o365spray.core.utils import Helper  # type: ignore
-from o365spray.core.utils import text_colors  # type: ignore
-from o365spray.core.utils import ThreadWriter  # type: ignore
-from o365spray.core.utils import Defaults, DefaultFiles  # type: ignore
-
-# Handlers
-from o365spray.core.handlers import Enumerator  # type: ignore
-from o365spray.core.handlers import Sprayer  # type: ignore
-from o365spray.core.handlers import Validator  # type: ignore
