+++ o365spray/core/handlers/enumerator/__init__.py
@@ -1,0 +1,1 @@
+from o365spray.core.handlers.enumerator.enumerate import enumerate
