+++ o365spray/core/handlers/sprayer/__init__.py
@@ -1,0 +1,1 @@
+from o365spray.core.handlers.sprayer.spray import spray
