+++ o365spray/core/handlers/validator/__init__.py
@@ -1,0 +1,1 @@
+from o365spray.core.handlers.validator.validate import validate
