+++ o365spray/core/utils/__init__.py
@@ -1,4 +1,8 @@
-from o365spray.core.utils.helper import Helper  # type: ignore
 from o365spray.core.utils.colors import text_colors  # type: ignore
+from o365spray.core.utils.defaults import (  # type: ignore
+    Defaults,
+    DefaultFiles,
+)
+from o365spray.core.utils.helper import Helper  # type: ignore
+from o365spray.core.utils.logger import init_logger  # type: ignore
 from o365spray.core.utils.writer import ThreadWriter  # type: ignore
-from o365spray.core.utils.defaults import Defaults, DefaultFiles  # type: ignore
