+++ o365spray/core/utils/defaults.py
@@ -1,5 +1,7 @@
 #!/usr/bin/env python3
 
+import sys
+import asyncio
 from datetime import datetime
 
 # Get the current time in YYMMDDHHMM format to append
@@ -34,35 +36,33 @@ class Defaults:
 
     # HTTP Header Configuration
     HTTP_HEADERS = {
-        "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10.14; rv:69.0) Gecko/20100101 Firefox/69.0",
-        "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8",
-        "Accept-Language": "en-US,en;q=0.5",
-        "Accept-Encoding": "gzip, deflate",
         "DNT": "1",
+        "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8",
         "Connection": "keep-alive",
+        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/104.0.0.0 Safari/537.36",
+        "Accept-Encoding": "gzip, deflate",
+        "Accept-Language": "en-US,en;q=0.5",
         "Upgrade-Insecure-Requests": "1",
     }
 
     # https://docs.microsoft.com/en-us/azure/active-directory/develop/reference-aadsts-error-codes
-    # https://gist.github.com/byt3bl33d3r/19a48fff8fdc34cc1dd1f1d2807e1b7f
-    # This will be used for both Autodiscover and Azure AD
+    # fmt: off
     AADSTS_CODES = {
-        "AADSTS50053": ["LOCKED", "Account locked"],
-        "AADSTS50055": ["EXPIRED_PASS", "Password expired"],
-        "AADSTS50057": ["DISABLED", "User disabled"],
-        "AADSTS50126": ["INVALID_CREDS", "Invalid username or password"],
-        "AADSTS50059": ["MISSING_TENANT", "Tenant for account doesn't exist"],
-        "AADSTS50128": ["INVALID_DOMAIN", "Tenant for account doesn't exist"],
-        "AADSTS50034": ["USER_NOT_FOUND", "User does not exist"],
-        "AADSTS50079": ["VALID_MFA", "Response indicates MFA (Microsoft)"],
-        "AADSTS50076": ["VALID_MFA", "Response indicates MFA (Microsoft)"],
-        "AADSTS50158": [
-            "SEC_CHAL",
-            "Response indicates conditional access (MFA: DUO or other)",
-        ],
+        "AADSTS50053":  ["LOCKED",           "Account locked"],
+        "AADSTS50055":  ["EXPIRED_PASS",     "Password expired"],
+        "AADSTS50057":  ["DISABLED",         "User disabled"],
+        "AADSTS50126":  ["INVALID_CREDS",    "Invalid username or password"],
+        "AADSTS50059":  ["MISSING_TENANT",   "Tenant for account doesn't exist"],
+        "AADSTS50128":  ["INVALID_DOMAIN",   "Tenant for account doesn't exist"],
+        "AADSTS50034":  ["USER_NOT_FOUND",   "User does not exist"],
+        "AADSTS50079":  ["VALID_MFA",        "Response indicates MFA (Microsoft)"],
+        "AADSTS50076":  ["VALID_MFA",        "Response indicates MFA (Microsoft)"],
+        "AADSTS50158":  ["SEC_CHAL",         "Response indicates conditional access (MFA: DUO or other)"],
         "AADSTS500011": ["INVALID_RESOURCE", "Invalid resource name"],
-        "AADSTS700016": ["INVALID_APPID", "Invalid application client ID"],
+        "AADSTS700016": ["INVALID_APPID",    "Invalid application client ID"],
+        "AADSTS53003":  ["VALID_CAP",        "Access blocked via Conditional Access Policies"],
     }
+    # fmt: on
 
     # List of valid AADSTS codes to check against
     VALID_AADSTS_CODES = [
@@ -72,6 +72,7 @@ class Defaults:
         "AADSTS50079",  # VALID_MFA
         "AADSTS50076",  # VALID_MFA
         "AADSTS50158",  # SEC_CHAL
+        "AADSTS53003",  # VALID_CAP
     ]
 
     # List of substrings that can be found when BasicAuth is blocked
@@ -80,3 +81,11 @@ class Defaults:
         "BasicAuthBlockStatus - Deny",
         "BlockBasicAuth - User blocked",
     ]
+
+    # Create a type handler for asyncio loops based on operating system
+    if sys.platform == "win32":
+        # windows
+        EventLoop = asyncio.windows_events.ProactorEventLoop
+    else:
+        # darwin/linux
+        EventLoop = asyncio.unix_events._UnixSelectorEventLoop
