+++ setup.cfg
@@ -21,6 +21,9 @@ classifiers =
 [options]
 packages = find:
 install_requires =
+    beautifulsoup4>=4.11.1
+    colorama>=0.4.5
+    lxml>=4.9.1
     requests>=2.22.0
 python_requires = >=3.6.1
